browser_automation_test-main/requirements.txt
urllib3==1.26.0
pandas==1.3.5
numpy==1.17.3
XlsxWriter==1.0.5
six==1.11.0
python_dateutil==2.7.3
Requests==2.26.0
opencv-python==4.2.0.32
scikit-image==0.17.2
pathlib==1.0.1
configparser==3.5.0
PyPubSub==3.3.0
argparse==1.4.0
aidea
